dev-requirements.txt
pip>=19.0.0
bumpversion>=0.5.3
wheel>=0.31.0
flake8==3.5.0
tox>=3.2.0
sphinx>0.3.2
sphinx_rtd_theme>=0.4.3
numpydoc>=1.1
coverage>=4.5.0
twine>=3.0
pytest==7.4.2
pytest-cov>=4.x
matplotlib>=3.7.1
ipython>7.10
